Python/ebook_crawl/requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
mysql-connector-python==8.0.17
pylint==2.4.2
python-telegram-bot==12.2.0
requests==2.18.4
virtualenv==16.7.5

